- unit of work, transactions

+++ Judge/Judge.Application/ViewModels/Submit/SubmitSolutionViewModel.cs
@@ -7,7 +7,7 @@ namespace Judge.Application.ViewModels.Submit
     public sealed class SubmitSolutionViewModel
     {
         public IEnumerable<LanguageViewModel> Languages { get; set; }
-        public long SelectedLanguage { get; set; }
+        public int SelectedLanguage { get; set; }
 
         [Required(ErrorMessage = "Выберите файл")]
         public HttpPostedFileBase File { get; set; }
